Remove section 12.3.7.2 (invariance of the meaning of names in a block)
This rule was never consistently implemented, and was removed in C# 6

Fixes #124

Note that I ran the section renumbering tool, which helped me find the references in statements.md. It made a number of unrelated changes which I've discarded, to keep this PR purely focused on the matter at hand.