tools: Separate consecutive notes/paragraphs
The Markdown parser considers them as a single span normally, so we
insert an extra "normal" paragraph with well-known text (before
Markdown parsing), so we can spot and remove it when converting the
parsed Markdown to Word.

Fixes #26.

This only works for notes at the top level. We have 29 notes that
occur within bullet points, but that's covered in a separate issue.