Fix deconstructing assignment
Deconstructing assignment is now fully integrated into simple assignment, and doesn't have its own section.

Deconstruction itself has been broken into an independent subsection of Expressions in anticipation of more deconstruction to come in future releases.

The evaluation order of a deconstructing assignment is now elegant and recursive. However, it still does not match the compiler behavior 100%. The evaluation order of the expressions directly occurring in the assignment expression is correct (except for what seems to be a compiler bug in a very specific case). However, in the case of nested left side tuples, some conversions and deconstructions may occur slightly later than my text implies. I've chosen to live with this difference since it is small, unobservable when the conversions and deconstructions are side-effect free (as they should be), and frankly beyond my capabilities to adequately overcome.