Add multi-file functionality (for a single project) (#735)
This fixes almost all examples where multiple files are required, example for VersioningOfConstantsAndStaticReadonlyFields1/2, which actually require multiple *projects*.

A few examples marked as requiring multiple files actually only needed additional files. It would be nice to be able to grab code from other examples to avoid maintenance notes, but that can be done later.

Note that the additional files generated have no template, and therefore no implicit using directives - hence the addition of "using System.Diagnostics;" in a couple of examples. This is *slightly* unfortunate, but not a significant issue - and the alternative of specifying template files would be much more work.

With this change in place, we will have 470 tested examples.